# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /admin/integrations/apps/toolkits/{slug}/tools:
    get:
      operationId: ComposioToolkitsController_findTools[0]
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - ComposioToolkits
      security:
        - bearer: []
  /admin/integrations/apps/toolkits/{slug}/tools/{tool_slug}:
    patch:
      operationId: ComposioToolkitsController_updateTool[0]
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
        - name: tool_slug
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - ComposioToolkits
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/integrations/apps/toolkits:
    get:
      operationId: OrgToolkitsController_listToolkits[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/toolkits/count:
    get:
      operationId: OrgToolkitsController_countToolkits[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/toolkits/{slug}/tools:
    get:
      operationId: OrgToolkitsController_listToolkitTools[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/toolkits/{slug}:
    get:
      operationId: OrgToolkitsController_getToolkit[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/toolkits/{slug}/enable:
    post:
      operationId: OrgToolkitsController_enableToolkit[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/toolkits/{slug}/disable:
    post:
      operationId: OrgToolkitsController_disableToolkit[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/tools/{tool_slug}:
    patch:
      operationId: OrgToolkitsController_updateToolPermission[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: tool_slug
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgToolPermissionDto"
      responses:
        "200":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/tools:
    get:
      operationId: OrgToolkitsController_listEnabledTools[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - OrgToolkits
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/triggers:
    get:
      operationId: ComposioTriggersController_list[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - ComposioTriggers
      security:
        - bearer: []
    post:
      operationId: ComposioTriggersController_create[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComposioTriggerDto"
      responses:
        "201":
          description: ""
      tags:
        - ComposioTriggers
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/triggers/{uuid}:
    patch:
      operationId: ComposioTriggersController_update[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComposioTriggerDto"
      responses:
        "200":
          description: ""
      tags:
        - ComposioTriggers
      security:
        - bearer: []
    delete:
      operationId: ComposioTriggersController_remove[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - ComposioTriggers
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/document-boards:
    get:
      operationId: DocumentBoardsController_findAll
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - DocumentBoards
      security:
        - bearer: []
    post:
      operationId: DocumentBoardsController_create
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBoardDto"
      responses:
        "201":
          description: ""
      tags:
        - DocumentBoards
      security:
        - bearer: []
  /organizations/{organization_uuid}/document-boards/{board_uuid}:
    get:
      operationId: DocumentBoardsController_findOne
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: board_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - DocumentBoards
      security:
        - bearer: []
    patch:
      operationId: DocumentBoardsController_update
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: board_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBoardDto"
      responses:
        "200":
          description: ""
      tags:
        - DocumentBoards
      security:
        - bearer: []
    delete:
      operationId: DocumentBoardsController_remove
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: board_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - DocumentBoards
      security:
        - bearer: []
  /organizations/{organization_uuid}/document-boards/{board_uuid}/items:
    post:
      operationId: DocumentBoardsController_addItem
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: board_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBoardItemDto"
      responses:
        "201":
          description: ""
      tags:
        - DocumentBoards
      security:
        - bearer: []
  /organizations/{organization_uuid}/document-boards/{board_uuid}/items/{item_uuid}:
    delete:
      operationId: DocumentBoardsController_removeItem
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: board_uuid
          required: true
          in: path
          schema:
            type: string
        - name: item_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - DocumentBoards
      security:
        - bearer: []
...
info:
  title: Cortex API
  description: >-
    REST API for Cortex — organizations, conversations, agents, integrations (Composio, MCP, OpenAPI, databases),
    documents, roles, and members. Authenticate via JWT Bearer token from POST /auth/email/login.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3001
    description: Local
  - url: "{{baseUrl}}"
    description: Environment variable
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/email/login
  schemas:
...
    UpdateOrgToolPermissionDto:
      type: object
      properties: {}
    CreateComposioTriggerDto:
      type: object
      properties: {}
    UpdateComposioTriggerDto:
      type: object
      properties: {}
    CreateDocumentBoardDto:
      type: object
      properties: {}
    UpdateDocumentBoardDto:
      type: object
      properties: {}
    AddBoardItemDto:
      type: object
      properties: {}
...
security:
  - bearer: []